Leave-one-out Cross Validation

This implements Leave-one-out cross-validation (LOO-CV) for trees of class aphylo and multiAphylo.

aphylo_cv(...)

# S3 method for formula
aphylo_cv(model, ...)

Arguments

...
Further arguments passed to the method.

model
As passed to aphylo_mcmc.

Value

An object of class aphylo_cv with the following components:

pred_out Out of sample prediction.

expected Expected annotations

call The call

ids Integer vector with the ids of the leafs used in the loo process.

Details

For each observation in the dataset (either a single gene if of class aphylo, or an entire tree if of class multiAphylo), we restimate the model removing the observation and use the parameter estimates to make a prediction on it. The prediction is done using the function predict.aphylo_estimates with argument loo = TRUE.