Objects of class aphylo_estimates

The model fitting of annotated phylogenetic trees can be done using either MLE via aphylo_mle() or MCMC via aphylo_mcmc(). This section describes the object of class aphylo_estimates that these functions generate and the post estimation methods/functions that can be used.

# S3 method for aphylo_estimates
print(x, ...)

# S3 method for aphylo_estimates
coef(object, ...)

# S3 method for aphylo_estimates
vcov(object, ...)

# S3 method for aphylo_estimates
plot(
  x,
  y = NULL,
  which.tree = 1L,
  ids = list(1:Ntip(x)[which.tree]),
  loo = TRUE,
  nsamples = 1L,
  ncores = 1L,
  centiles = c(0.025, 0.5, 0.975),
  cl = NULL,
  ...
)

Arguments

x, object
Depending of the method, an object of class aphylo_estimates.

...
Further arguments passed to the corresponding method.

y
Ignored.

which.tree
Integer scalar. Which tree to plot.

ids, nsamples, ncores, centiles, cl
passed to predict.aphylo_estimates()

loo
Logical scalar. When loo = TRUE, predictions are preformed similar to what a leave-one-out cross-validation scheme would be done (see predict.aphylo_estimates).

Value

Objects of class aphylo_estimates are a list withh the following elements:

par

A numeric vector of length 5 with the solution.

hist

A numeric matrix of size counts*5 with the solution path (length 2 if used optim as the intermediate steps are not available to the user). In the case of aphylo_mcmc, hist is an object of class coda::mcmc.list().

ll

A numeric scalar with the value of fun(par, dat). The value of the log likelihood.

counts

Integer scalar number of steps/batch performed.

convergence

Integer scalar. Equal to 0 if optim converged. See optim.

message

Character scalar. See optim.

fun

A function (the objective function).

priors

If specified, the function priors passed to the method.

dat

The data dat provided to the function.

par0

A numeric vector of length 5 with the initial parameters.

method

Character scalar with the name of the method used.

varcovar

A matrix of size 5*5. The estimated covariance matrix.

The plot method for aphylo_estimates returns the selected tree (which.tree) with predicted annotations, also of class aphylo.
Details

The plot method for the object of class aphylo_estimates plots the original tree with the predicted annotations.