Approximation of Geodesic distances using Matrix Powers

Given an adjacency matrix \(A\), the geodesic can be approximated using its powers, since each \((i,j)\) element of \(A^p\) corresponds to the number of \(p\) length steps between nodes \(i\) and \(j\).

approx_geodesic(edges, nsteps = 5000L, undirected = TRUE, warn = FALSE)

Arguments

edges
Integer matrix with two columns. An edgelist where each row is in the form of \code(parent, offspring).

nsteps
Integer scalar. Number of maximum steps for the approximation.

undirected
Logical scalar. When TRUE (default), the edgelist is treated as undirected (see details).

warn
Logical scalar. When TRUE shows a warning after no further steps are needed.

Value

A square matrix of size \(n\) with the shortest path between each pair of nodes.

Details

When undirected = TRUE, the function extends edges such that edges = rbind(edges, edges[,2:1]).

References

This is a modified version of the function of the same name in the R package netdiffuseR.