Default priors for aphylo_mcmc

Convenient wrappers to be used with the aphylo estimation methods.

bprior(shape1 = 1, shape2 = 9, ...)

uprior()

Arguments

shape1, shape2, ...

Arguments passed to stats::dbeta

Value

In the case of bprior, a wrapper of the function stats::dbeta. uprior returns a function function(p) 1 (the uniform prior)

Examples

bprior(1, 9)
#> function (p) 
#> {
#>     stats::dbeta(p, shape1 = shape1, shape2 = shape2, ...)
#> }
#> <bytecode: 0x55897e9ad3b8>
#> <environment: 0x55897e9acdd0>
uprior()
#> function (p) 
#> 1
#> <bytecode: 0x55897e0d2c60>
#> <environment: 0x558979af5268>