Pointer to pruner

Creates an external pointer to an object of class aphylo_pruner. This is mostly used to compute the model's likelihood function faster by reusing underlying C++ class objects to store probability matrices and data. This is intended for internal use only.

dist2root(ptr)

get_postorder(ptr)

new_aphylo_pruner(x, ...)

Arguments

ptr
An object of class aphylo_pruner.

x
An object of class aphylo or multiAphylo.

...
Further arguments passed to the method

Value

dist2root: An integer vector with the number of steps from each node (internal or not) to the root node.

get_postorder: An integer vector with the postorder sequence for pruning the tree (indexed from 0).

The function new_aphylo_pruner returns an object of class aphylo_pruner or multiAphylo_pruner, depending on the class of x.

Details

The underlying implementation of the pruning function is based on the pruner C++ library that implements Felsenstein's tree pruning algorithm. See https://github.com/USCbiostats/pruner.

Examples

set.seed(1)
x  <- raphylo(10) 
pruner <- new_aphylo_pruner(x)

# Computing loglike
LogLike(
  pruner,
  psi  = c(.10, .20),
  mu_d = c(.90, .80),
  mu_s = c(.10, .05),
  Pi   = .05,
  eta  = c(.90, .80)
  )

#> $Pr
#> $Pr[[1]]
#>              [,1]         [,2]
#>  [1,] 0.080000000 6.400000e-01
#>  [2,] 0.810000000 1.800000e-01
#>  [3,] 0.080000000 6.400000e-01
#>  [4,] 0.810000000 1.800000e-01
#>  [5,] 0.080000000 6.400000e-01
#>  [6,] 0.810000000 1.800000e-01
#>  [7,] 0.810000000 1.800000e-01
#>  [8,] 0.810000000 1.800000e-01
#>  [9,] 0.810000000 1.800000e-01
#> [10,] 0.080000000 6.400000e-01
#> [11,] 0.002512323 2.849105e-05
#> [12,] 0.015483128 3.271910e-04
#> [13,] 0.197940493 1.722348e-02
#> [14,] 0.033606163 3.123736e-03
#> [15,] 0.068908660 4.308241e-02
#> [16,] 0.558009000 4.473225e-02
#> [17,] 0.558009000 4.473225e-02
#> [18,] 0.558009000 4.473225e-02
#> [19,] 0.018496000 3.745440e-01
#> 
#> 
#> $ll
#> [1] -6.037244
#>