Reads PANTHER db trees

The PANTHER Project handles a modified version of newick tree files which, besides of the tree structure, includes the type of node and ancestor labels. This function is a wrapper of ape::read.tree().

read_panther(x, tree.reader = ape::read.tree, ...)

read.panther(x, tree.reader = ape::read.tree, ...)

Arguments

x
Character scalar. Full path to the panther file.

tree.reader
Function that will be used to read the tree file. It can be either ape::read.tree or rncl::read_newick_phylo.

...
Further arguments passed to ape::read.tree().

Value

A list consisting of a data.frame and a phylo object. The data.frame has the following columns:

branch_length

Numeric vector. Length of the branch to its parent node.

type

Character vector. Can be either "S" (speciation), "D" (duplication), or "T" (horizontal transfer).

ancestor

Character vector. Name of the ancestor.

The nodeids can be identified using the rownames.
See also

Examples

path <- system.file("tree.tree", package="aphylo")
read_panther(path)

#> $tree
#> 
#> Phylogenetic tree with 145 tips and 107 internal nodes.
#> 
#> Tip labels:
#>   AN5:MONBE|Gene=28576|UniProtKB=A9V8K6, AN7:SCHPO|PomBase=SPAC25B8.12c|UniProtKB=Q9UTA6, AN8:SCHPO|PomBase=SPBC215.10|UniProtKB=O94314, AN11:ENTHI|EnsemblGenome=EHI_168190|UniProtKB=C4M4Q5, AN12:ENTHI|EnsemblGenome=EHI_151930|UniProtKB=C4LSN2, AN13:ENTHI|EnsemblGenome=EHI_149870|UniProtKB=C4M9D2, ...
#> Node labels:
#>   AN0, AN1, AN2, AN3, AN4, AN6, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> $internal_nodes_annotations
#>       branch_length type                ancestor duplication
#> AN0              NA    S                    LUCA       FALSE
#> AN1           0.057    S       Archaea-Eukaryota       FALSE
#> AN2           0.244    S               Eukaryota       FALSE
#> AN3           0.436    S                Unikonts       FALSE
#> AN4           0.417    S            Opisthokonts       FALSE
#> AN6           0.684    D                    <NA>        TRUE
#> AN9           0.790    S               Amoebozoa       FALSE
#> AN10          0.183    D                    <NA>        TRUE
#> AN15          0.570    S           Dictyostelium       FALSE
#> AN18          0.519    S Alveolata-Stramenopiles       FALSE
#> AN19          0.427    S               Alveolata       FALSE
#> AN20          0.353    D                    <NA>        TRUE
#> AN25          0.312    D                    <NA>        TRUE
#> AN28          0.544    S           Viridiplantae       FALSE
#> AN30          0.454    S             Embryophyta       FALSE
#> AN31          0.216    S           Magnoliophyta       FALSE
#> AN32          0.194    S                 Poaceae       FALSE
#> AN34          0.048    D                    <NA>        TRUE
#> AN35          0.667    S               BEP_clade       FALSE
#> AN36          0.100    D                    <NA>        TRUE
#> AN40          0.049    S               BEP_clade       FALSE
#> AN41          0.027    D                    <NA>        TRUE
#> AN45          0.137    S            Pentapetalae       FALSE
#> AN47          0.090    S                  rosids       FALSE
#> AN50          0.399    S                  fabids       FALSE
#> AN51          0.094    D                    <NA>        TRUE
#> AN54          0.039    D                    <NA>        TRUE
#> AN57          0.094    D                    <NA>        TRUE
#> AN60          0.653    S                 Archaea       FALSE
#> AN61          0.304    S            Thermoprotei       FALSE
#> AN62          0.376    D                    <NA>        TRUE
#> AN66          0.452    S           Euryarchaeota       FALSE
#> AN70          0.057    D                    <NA>        TRUE
#> AN71          0.958    S              Eubacteria       FALSE
#> AN72          0.209    S                 Bacilli       FALSE
#> AN74          0.339    S              Bacillales       FALSE
#> AN77          0.426    D                    <NA>        TRUE
#> AN78          0.404    S      Enterobacteriaceae       FALSE
#> AN79          0.027    D                    <NA>        TRUE
#> AN83          0.351    S      Enterobacteriaceae       FALSE
#> AN86          1.201    S              Eubacteria       FALSE
#> AN89          0.435    T                    <NA>        TRUE
#> AN90          0.294    S        Trypanosomatidae       FALSE
#> AN93          0.399    S        Trypanosomatidae       FALSE
#> AN94          0.183    D                    <NA>        TRUE
#> AN98          0.320    D                    <NA>        TRUE
#> AN101         0.231    T                    <NA>        TRUE
#> AN104         0.321    S     Gammaproteobacteria       FALSE
#> AN107         0.226    D                    <NA>        TRUE
#> AN108         0.594    S      Enterobacteriaceae       FALSE
#> AN110         0.468    D                    <NA>        TRUE
#> AN114         0.413    S      Enterobacteriaceae       FALSE
#> AN118         0.811    S              Eubacteria       FALSE
#> AN119         0.143    D                    <NA>        TRUE
#> AN120         0.422    S                 Bacilli       FALSE
#> AN122         0.461    S              Bacillales       FALSE
#> AN123         0.084    D                    <NA>        TRUE
#> AN127         0.193    S                Bacillus       FALSE
#> AN128         0.077    D                    <NA>        TRUE
#> AN131         0.056    D                    <NA>        TRUE
#> AN135         0.285    D                    <NA>        TRUE
#> AN138         0.677    S              Eubacteria       FALSE
#> AN139         0.608    D                    <NA>        TRUE
#> AN140         0.366    S              Firmicutes       FALSE
#> AN142         0.341    S              Bacillales       FALSE
#> AN143         0.340    D                    <NA>        TRUE
#> AN147         0.436    S                Bacillus       FALSE
#> AN148         0.113    D                    <NA>        TRUE
#> AN151         0.125    D                    <NA>        TRUE
#> AN154         0.389    S              Firmicutes       FALSE
#> AN155         0.043    D                    <NA>        TRUE
#> AN159         0.029    D                    <NA>        TRUE
#> AN160         0.673    S                 Bacilli       FALSE
#> AN162         0.543    S              Bacillales       FALSE
#> AN164         0.338    D                    <NA>        TRUE
#> AN167         0.221    S                Bacillus       FALSE
#> AN170         0.710    S                 Bacilli       FALSE
#> AN172         0.254    D                    <NA>        TRUE
#> AN175         0.391    S                 Bacilli       FALSE
#> AN176         0.175    D                    <NA>        TRUE
#> AN179         0.321    S              Bacillales       FALSE
#> AN181         0.089    D                    <NA>        TRUE
#> AN185         0.348    D                    <NA>        TRUE
#> AN186         0.631    S                Bacillus       FALSE
#> AN189         0.325    S                Bacillus       FALSE
#> AN190         0.092    D                    <NA>        TRUE
#> AN193         0.051    D                    <NA>        TRUE
#> AN197         0.752    D                    <NA>        TRUE
#> AN198         0.292    S         Actinomycetales       FALSE
#> AN199         0.314    D                    <NA>        TRUE
#> AN202         0.519    D                    <NA>        TRUE
#> AN207         1.294    D                    <NA>        TRUE
#> AN211         0.922    D                    <NA>        TRUE
#> AN214         0.651    S           Cyanobacteria       FALSE
#> AN215         0.193    D                    <NA>        TRUE
#> AN220         0.878    D                    <NA>        TRUE
#> AN224         1.050    D                    <NA>        TRUE
#> AN227         0.479    S          Proteobacteria       FALSE
#> AN229         0.348    D                    <NA>        TRUE
#> AN230         0.444    S     Gammaproteobacteria       FALSE
#> AN232         0.597    S      Enterobacteriaceae       FALSE
#> AN233         0.064    D                    <NA>        TRUE
#> AN236         0.354    D                    <NA>        TRUE
#> AN241         0.432    S     Gammaproteobacteria       FALSE
#> AN244         0.294    D                    <NA>        TRUE
#> AN245         1.043    S      Enterobacteriaceae       FALSE
#> AN248         0.619    S      Enterobacteriaceae       FALSE
#>