Visualize predictions

Visualize predictions

# S3 method for aphylo_prediction_score
plot(x, y = NULL,
  main = "Prediction Accuracy: Observed versus predicted values",
  main.colorkey = "Probability of Functional Annotation",
  which.fun = seq_len(ncol(x$expected)), include.labels = NULL,
  labels.col = "black", leafs_only = TRUE, ...)

Arguments

x
An object of class aphylo_prediction_score.

y
Ignored.

main
Passed to title.

main.colorkey
Character scalar. Title of the colorkey (optional).

which.fun
Integer vector. Which function to plot.

include.labels
Logical scalar. When TRUE, draws nice labels at each slice which by default are specified as the rownames of x$expected. This is mostly useful when the number of predictions is small.

labels.col
Character scalar. Color of the labels.

leafs_only
Logical. When TRUE (default) only plots the leaf nodes.

...
Ignored

Details

If include.labels = NULL and ncol(x$expected) > 40, then include.labels=FALSE by default.