Calculate prediction score (quality of prediction)

Calculate prediction score (quality of prediction)

prediction_score(x, expected, alpha = NULL, W = NULL)

# S3 method for default
prediction_score(x, expected, alpha = NULL, W = NULL)

# S3 method for aphylo_estimates
prediction_score(x, expected, alpha = NULL,
  W = NULL, ...)

Arguments

x
An object of class aphylo_estimates or a numeric matrix.

expected
Integer vector of length \(n\). Expected values (either 0 or 1).

alpha
Numeric scalar. Prior belief of the parameter of the bernoulli distribution used to compute the random imputation score.

W
A square matrix. Must have as many rows as genes in expected.

...
Further arguments passed to predict.aphylo_estimates

Details

In the case of prediction_score, ... are passed to predict.aphylo_estimates.

In the case of the method for aphylo estimates, the function takes as a reference using alpha equal to the proportion of observed tip annotations that are equal to 1, this is:

mean(x$dat$tip.annotation[x$dat$tip.annotation != 9L], na.rm = TRUE)

Examples

# Example with prediction_score ---------------------------------------------
set.seed(1312)
ap  <- raphylo(10, P = 1, Pi=.2, mu=c(.05,.02))
ans <- aphylo_mcmc(
  ap ~ mu + eta + psi + Pi,
  control = list(nsteps=2e4, thin=100),
  priors  = function(x) dbeta(x, 1, 30)
  )

#> No parameters were specified. Default will be used instead.

#> Warning: A single initial point has been passed via `initial`: c(0.1, 0.1, 0.1, 0.1, 0.9, 0.9, 0.1). The values will be recycled.

#> Convergence has been reached with 16000 steps (55 final count of observations).

pr <- prediction_score(ans)
with(pr, cbind(Expected = expected, Predicted = predicted))

#>    fun0000      fun0000
#> 1        0 8.593588e-04
#> 2        0 8.593128e-04
#> 3        0 8.593588e-04
#> 4        0 8.593588e-04
#> 5        0 8.593588e-04
#> 6        0 8.607876e-04
#> 7        0 8.340264e-04
#> 8        0 8.593128e-04
#> 9        0 8.607876e-04
#> 10       0 8.362862e-04
#> 11       0 3.073605e-05
#> 12       0 5.198681e-05
#> 13       0 3.128361e-05
#> 14       0 3.047520e-05
#> 15       0 1.017682e-04
#> 16       0 1.018586e-04
#> 17       0 5.643579e-05
#> 18       0 1.046716e-04
#> 19       0 1.018586e-04